requirements/_base.txt
argparse==1.4.0
behave==1.2.5
decorator==4.0.2
django-appconf==0.6
django-ckeditor==4.5.0
django-compressor==1.4
django-extensions==1.5.7
django-registration-redux==1.2
Django==1.8.4
djangocms-admin-style==0.2.8
djangorestframework==3.2.4
ipdb==0.8.1
ipython-genutils==0.1.0
ipython==4.0.0
path.py==7.6.1
pexpect==3.3
pickleshare==0.5
python-dateutil==2.4.2
python-memcached==1.57
simplegeneric==0.8.1
six==1.7.3
selenium==2.47.3
sqlparse==0.1.16
traitlets==4.0.0
xlwt==1.0.0
djangoajax==2.2.15
django-geojson==2.8.1
django-leaflet==0.17.0
jsonfield
